refactor
refactor validators, add core folder
