+++ src/core/enums/days.enum.ts
@@ -1,0 +1,12 @@
+export enum Days {
+  MONDAY,
+  TUESDAY,
+  WEDNESDAY,
+  THURSDAY,
+  FRIDAY,
+  SATURDAY,
+  SUNDAY
+}
+
+export const WEEKDAYS = [ Days.MONDAY, Days.TUESDAY, Days.WEDNESDAY, Days.THURSDAY, Days.FRIDAY ];
+export const WEEKENDS = [ Days.SATURDAY, Days.SUNDAY ];
