+++ src/helpers/date.ts
@@ -1,0 +1,7 @@
+import * as moment from 'moment';
+
+export namespace Date {
+  export const DATE_PATTERN = 'YYYY-MM-DD';
+
+  export const formatDate = (value, pattern = DATE_PATTERN) => moment(value, pattern).valueOf();
+}
